core/src: Add half math fns
  - Add half math fns to non-public header
  - Move floating point wrapper to non-public header
